Fix type errors (#365)
* fix most ts errors

* Update Makefile with type-check

* Change window.location.replace into navigate from gatsby

* import right types and use them instead of 'any' :tada:
