+++ src/custom.d.ts
@@ -13,14 +13,6 @@ declare module "*.js" {
   export default content;
 }
 
-import MyInterface from "./MyInterface";
-
-declare global {
-  interface Window {
-    docsearch: any;
-  }
-}
-
 declare module "prism-react-renderer";
 declare module "react-modal-hook";
 declare module "get-video-id";
