+++ src/window.d.ts
@@ -1,0 +1,6 @@
+declare global {
+  interface Window {
+    docsearch: any;
+  }
+}
+export {};
